requirements.txt
Django==2.0.7
django-tinymce==2.7.0
Pillow==7.1.1
pytz==2018.7
pymysql
